fix: scheduler fix + mock data
